+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         <meta-data android:name="DOMAIN_PACKAGE_NAME" android:value="com.example.posmedicine.models" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -30,28 +30,28 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".CreateUnitActivity" />
+        <activity android:name=".activities.CreateUnitActivity" />
         <!-- <activity android:name=".EditUnitActivity" /> -->
         <activity
-            android:name=".MedicineActivity"
+            android:name=".activities.MedicineActivity"
             android:label="@string/title_activity_medicine"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".UnitActivity"
+            android:name=".activities.UnitActivity"
             android:label="@string/title_activity_unit"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".AppointmentActivity"
+            android:name=".activities.AppointmentActivity"
             android:label="@string/title_activity_appointment"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".AppointmentDetailActivity" />
-        <activity android:name=".CreateMedicineActivity" />
-        <activity android:name=".EditMedicineActivity" />
-        <activity android:name=".EditUnitActivity" />
-        <activity android:name=".PharmacyChasierActivity" />
-        <activity android:name=".CashierTransactionActivity" />
-        <activity android:name=".CashierTransactionDetailActivity" />
-        <activity android:name=".CreateAppointmentActivity"></activity>
+        <activity android:name=".activities.AppointmentDetailActivity" />
+        <activity android:name=".activities.CreateMedicineActivity" />
+        <activity android:name=".activities.EditMedicineActivity" />
+        <activity android:name=".activities.EditUnitActivity" />
+        <activity android:name=".activities.PharmacyChasierActivity" />
+        <activity android:name=".activities.CashierTransactionActivity" />
+        <activity android:name=".activities.CashierTransactionDetailActivity" />
+        <activity android:name=".activities.CreateAppointmentActivity"></activity>
     </application>
 
 </manifest>
