+++ app/src/main/java/com/example/posmedicine/Adapter/MedicineAdapter.java
@@ -5,7 +5,6 @@ import android.os.Bundle;
 import android.os.Parcelable;
 import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,8 +12,8 @@ import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.posmedicine.EditMedicineActivity;
-import com.example.posmedicine.MedicineActivity;
+import com.example.posmedicine.activities.EditMedicineActivity;
+import com.example.posmedicine.activities.MedicineActivity;
 import com.example.posmedicine.R;
 import com.example.posmedicine.models.Medicine;
 import com.example.posmedicine.models.Unit;
