+++ app/src/main/java/com/example/posmedicine/Adapter/UnitAdapter.java
@@ -13,11 +13,10 @@ import android.view.LayoutInflater;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
-//import com.example.posmedicine.EditUnitActivity;
-import com.example.posmedicine.EditUnitActivity;
+import com.example.posmedicine.activities.EditUnitActivity;
 import com.example.posmedicine.R;
 
-import com.example.posmedicine.UnitActivity;
+import com.example.posmedicine.activities.UnitActivity;
 
 import com.example.posmedicine.models.Unit;
 import com.example.posmedicine.models.response.UnitResponse;
