+++ app/src/main/java/com/example/posmedicine/Adapter/local/LocalAppointmentAdapter.java
@@ -10,11 +10,9 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.example.posmedicine.AppointmentActivity;
+import com.example.posmedicine.activities.AppointmentActivity;
 import com.example.posmedicine.FontManager;
 import com.example.posmedicine.R;
-import com.example.posmedicine.models.Appointment;
-import com.example.posmedicine.models.Doctor;
 import com.example.posmedicine.models.local.LocalAppointment;
 
 import java.util.List;
