+++ app/src/main/java/com/example/posmedicine/activities/AppointmentDetailActivity.java
@@ -1,9 +1,11 @@
-package com.example.posmedicine;
+package com.example.posmedicine.activities;
 
 import android.graphics.Typeface;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.widget.TextView;
+
+import com.example.posmedicine.FontManager;
+import com.example.posmedicine.R;
 
 public class AppointmentDetailActivity extends AppCompatActivity {
