+++ app/src/main/res/layout/activity_appointment_detail.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.AppointmentDetailActivity">
+    tools:context="com.example.posmedicine.activities.AppointmentDetailActivity">
 
     <GridLayout
         android:layout_width="match_parent"
