+++ app/src/main/res/layout/activity_cashier_transaction.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.CashierTransactionActivity">
+    tools:context="com.example.posmedicine.activities.CashierTransactionActivity">
 
     <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
