+++ app/src/main/res/layout/activity_cashier_transaction_detail.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.CashierTransactionDetailActivity">
+    tools:context="com.example.posmedicine.activities.CashierTransactionDetailActivity">
     <LinearLayout
         android:id="@+id/headerContainer"
         android:layout_height="wrap_content"
