+++ app/src/main/res/layout/activity_create_medicine.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.CreateMedicineActivity">
+    tools:context="com.example.posmedicine.activities.CreateMedicineActivity">
 
     <AutoCompleteTextView
         android:hint="Medicine Unit"
