+++ app/src/main/res/layout/activity_create_unit.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.CreateUnitActivity">
+    tools:context="com.example.posmedicine.activities.CreateUnitActivity">
 
     <EditText
         android:layout_width="match_parent"
