+++ app/src/main/res/layout/activity_edit_medicine.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.EditMedicineActivity">
+    tools:context="com.example.posmedicine.activities.EditMedicineActivity">
 
     <AutoCompleteTextView
         android:hint="Medicine Unit"
