+++ app/src/main/res/layout/activity_edit_unit.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.EditUnitActivity">
+    tools:context="com.example.posmedicine.activities.EditUnitActivity">
 
     <EditText
         android:layout_width="match_parent"
