+++ app/src/main/res/layout/activity_medicine.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.posmedicine.MedicineActivity">
+    tools:context="com.example.posmedicine.activities.MedicineActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
