+++ app/src/main/res/layout/activity_pharmacy_chasier.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.posmedicine.PharmacyChasierActivity">
+    tools:context="com.example.posmedicine.activities.PharmacyChasierActivity">
 
 
     <AutoCompleteTextView
