+++ app/src/main/res/layout/activity_unit.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.posmedicine.UnitActivity">
+    tools:context="com.example.posmedicine.activities.UnitActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
