+++ app/src/main/res/layout/content_appointment.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.posmedicine.AppointmentActivity"
+    tools:context="com.example.posmedicine.activities.AppointmentActivity"
     tools:showIn="@layout/activity_appointment">
 
     <android.support.v7.widget.RecyclerView
