+++ app/src/main/res/layout/content_unit.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.posmedicine.UnitActivity"
+    tools:context="com.example.posmedicine.activities.UnitActivity"
     tools:showIn="@layout/activity_unit">
 
     <android.support.v7.widget.RecyclerView
